Truong/function app all

+++ app/src/main/java/com/example/petify/data/server/enitities/ProductModel.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class ProductModel(
     @field:SerializedName("_id") var id: String = "",
     @field:SerializedName("supplier_id") var supplierId: String = "",
-    @field:SerializedName("price") var price: Int = 0,
+    @field:SerializedName("price") var price: Double = 0.0,
     @field:SerializedName("date") var date: String = "",
     @field:SerializedName("expiry_Date") var expiryDate: String = "",
     @field:SerializedName("quantity") var quantity: Int = 0,
